Move clonning logic back to the mutation engine class
In retrospect, this is where it belongs, Arbor itself does not care how nodes are cloned, the engine is how we can configure it to be mutable vs immutable.
